fix: Fixed a bug in the countKeywords method, and made changes to the tests.
